Add ElementIsFinished property to NameSpaceElement
Added a new boolean property 'ElementIsFinished' to the 'NameSpaceElement' class. This property is used to determine if the element has been fully processed or not. Updated the 'AddParameter' method in 'NameSpaceElement' to check this property before adding parameters. Also, updated the 'TokenParser' class to set this property when a TypeDeclarationKeyword token is encountered.