Add PropertySetterTranslation class and update PropertySyntaxElement
Added a new class, PropertySetterTranslation, to handle the translation of Oxygene property setters to C#. This class uses OpenAI API for translation. Also updated the PropertySyntaxElement class to use this new translator for setter code. The changes ensure that both getter and setter codes are translated correctly from Oxygene to C#.