Add notify keyword detection in Sharpener
This update introduces the ability to detect the use of the "notify" keyword in files parsed by Sharpener. A new property, IsNotifyKeywordUsedInFile, has been added to the Document class to track this information.

The NameSpaceElement and PropertySyntaxElement classes have been updated to react accordingly when this keyword is detected. In particular, if a file uses "notify", then "PropertyChanged" will be added to Usings in NameSpaceElement.

Additionally, changes have been made to how PropertySyntaxElement handles tokens and generates code nodes based on whether or not "notify" is used in a file. The TokenParser class has also been updated for better handling of NotifyPropertyChangedImplementation token type.

Lastly, a LastUsedProperty property was added to Document class for tracking last used property syntax element which aids in applying notify pattern correctly.