Added Constructors and Regions to TODO list
In the Sharpener project, two new tasks have been added to the TODO.txt file: 'Constructors' and 'Regions'. Additionally, in Tokenizer.cs, a condition has been inserted to check for lines containing '$region' or '$endregion', incrementing the index if found. This is a temporary fix for handling regions which will need a more proper solution later.