Add EventsAndInterfacesExample and update Sharpener
Added a new file, EventsAndInterfacesExample.pas, which includes the implementation of an interface with a property, event, and method. Also created a class that implements this interface.

Updated Document.cs in SyntaxTree to move some properties to the top of the class for better visibility and added JsonIgnore and IgnoreDataMember attributes to these properties.

Refactored PropertySyntaxElement.cs by removing IsNullable property and moving its logic to PropertySyntaxElementBase.cs. This change also involved updating how tokens are handled in WithToken method.

Added a new file, PropertySyntaxInInterfaceElement.cs, which handles property syntax within interfaces. It includes token handling for getter/setter keywords and semicolon token.